+++ api/src/directory/controllers/Resource/__init__.py
@@ -1,0 +1,5 @@
+"""
+`Resource` Controllers module. Controllers permit the addition,
+modification, deletion, fetching, and processing of data. `Resource`s
+provide the main content for `BI Portal`.
+"""