+++ api/src/directory/controllers/Resource/search/__init__.py
@@ -1,0 +1,5 @@
+"""
+`Resource` search controller module. Controller provides
+the ability to perform a specialized search for `Resource`s
+against those in the database.
+"""