+++ api/src/directory/controllers/Resource/tests/Resource/mutation/__init__.py
@@ -1,0 +1,5 @@
+"""
+Mutation test module. These tests validate the functionality
+of controllers that create, modify or delete `Resource` data
+from the database.
+"""