+++ api/src/directory/controllers/Resource/tests/Resource/query/read/__init__.py
@@ -1,0 +1,4 @@
+"""
+Read test module. These tests validate the functionality
+of controllers that fetch `Resource` data from the database.
+"""