+++ CHANGELOG.md
@@ -18,10 +18,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -
 
-## [0.1.0] - 2024-8-22
+## [0.1.0] - 2024-8-31
 
 ### Added
 
 -   Forked ADAS Django React.
 -   Implemented Django apps & models for portal, directory, analytics, preferences, and request.
 -   Initialized BI Portal models.
+
+## [0.1.0] - 2024-9-30
+
+### Added
+
+-   Implemented frontend API state management.
+-   Implemented API Django view stubs.
+-   Implemented directory.resource.search.